requirements.txt
flask
opencv-python
numpy
pillow
arabic-reshaper
python-bidi
ffmpeg-python